+++ code/jvm/src/main/kotlin/pt/isel/ps/qq/controllers/GuestController.kt
@@ -1,14 +1,11 @@
 package pt.isel.ps.qq.controllers
 
-import org.springframework.http.HttpStatus
+
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
-import pt.isel.ps.qq.data.dto.ErrorDto
-import pt.isel.ps.qq.data.dto.UserDto
-import pt.isel.ps.qq.data.dto.UserTokenDto
 import pt.isel.ps.qq.data.dto.input.*
 import pt.isel.ps.qq.exceptions.AlreadyExistsException
 import pt.isel.ps.qq.exceptions.InvalidTokenException
@@ -20,9 +17,7 @@ import pt.isel.ps.qq.service.GuestService
 @RequestMapping("/api/web/v1.0/guest")
 class GuestController(
     private val service: GuestService
-) {
-
-    companion object {
+) { companion object {
         const val USER_DISPLAY_NAME_NOT_NULL = "displayName required"
         const val USER_USERNAME_ALREADY_EXISTS = "username already exists"
         const val USER_INVALID_TOKEN = "invalid token"
@@ -35,7 +30,7 @@ class GuestController(
             ResponseEntity.ok().body(service.register(input))
         } catch(e: AlreadyExistsException) {
             ResponseEntity.badRequest().body(
-                ErrorDto(code = HttpStatus.BAD_REQUEST.value(), reason = USER_USERNAME_ALREADY_EXISTS)
+                USER_USERNAME_ALREADY_EXISTS
             )
         }
     }
@@ -45,9 +40,9 @@ class GuestController(
         return try {
             ResponseEntity.ok().body(service.requestLogin(userName))
         } catch(e: InvalidTokenException) {
-            ResponseEntity.badRequest().body(ErrorDto(code = HttpStatus.BAD_REQUEST.value(), USER_INVALID_TOKEN))
+            ResponseEntity.badRequest().body(USER_INVALID_TOKEN)
         } catch(e: UserNotFoundException) {
-            ResponseEntity.badRequest().body(ErrorDto(code = HttpStatus.BAD_REQUEST.value(), USER_NOT_REGISTERED))
+            ResponseEntity.badRequest().body(USER_NOT_REGISTERED)
         }
 
     }
@@ -57,9 +52,9 @@ class GuestController(
         return try {
             ResponseEntity.ok().body(service.logmein(input))
         } catch(e: InvalidTokenException) {
-            ResponseEntity.badRequest().body(ErrorDto(code = HttpStatus.BAD_REQUEST.value(), USER_INVALID_TOKEN))
+            ResponseEntity.badRequest().body(USER_INVALID_TOKEN)
         } catch(e: UserNotFoundException) {
-            ResponseEntity.badRequest().body(ErrorDto(code = HttpStatus.BAD_REQUEST.value(), USER_NOT_REGISTERED))
+            ResponseEntity.badRequest().body(USER_NOT_REGISTERED)
         }
     }
