--- code/jvm/src/main/kotlin/pt/isel/ps/qq/data/DataClasses.kt
@@ -1,17 +1,0 @@
-package pt.isel.ps.qq.data
-
-data class JoinSession(
-    val guestId: String,
-    val sessionId: String,
-    val location: String? = null
-)
-
-data class GiveAnswer(
-    val sessionId: String,
-    val quizId: String,
-    val guestId: String,
-    val questionType: String,
-    val questionNumber: Int,
-    val answerNumber: Int? = null,
-    val answerText: String? = null
-)
