--- code/jvm/src/main/kotlin/pt/isel/ps/qq/data/User.kt
@@ -1,12 +1,0 @@
-package pt.isel.ps.qq.data
-
-data class User(
-    val userName: String,
-    var displayName: String,
-    val id: String,
-    val idExpiredDate: Long? = null
-) {
-    constructor(user: User, expiredDate: Long): this(user.userName, user.displayName, user.id, expiredDate)
-
-    constructor(user: User, uid: String, expiredDate: Long): this(user.userName, user.displayName, uid, expiredDate)
-}
