--- code/jvm/src/main/kotlin/pt/isel/ps/qq/data/dto/ErrorDto.kt
@@ -1,6 +1,0 @@
-package pt.isel.ps.qq.data.dto
-
-data class ErrorDto(
-    val code: Int,
-    val reason: String
-)
