--- code/jvm/src/main/kotlin/pt/isel/ps/qq/data/dto/UserDto.kt
@@ -1,6 +1,0 @@
-package pt.isel.ps.qq.data.dto
-
-data class UserDto(
-    val userName: String,
-    val displayName: String? = null
-)
