+++ code/jvm/src/main/kotlin/pt/isel/ps/qq/data/elasticdocs/UserDoc.kt
@@ -1,4 +1,4 @@
-package pt.isel.ps.qq.repositories.elasticdocs
+package pt.isel.ps.qq.data.elasticdocs
 import org.springframework.data.annotation.Id
 import org.springframework.data.elasticsearch.annotations.Document
 
@@ -7,7 +7,7 @@ data class UserDoc(
     @Id
     val userName: String,
     val displayName: String? = null,
-    val loginToken: String?,
+    val loginToken: String,
     val tokenExpireDate: Long? = null,
     val status: String? = null,
     val templates: List<String> = emptyList(),
