+++ code/jvm/src/main/kotlin/pt/isel/ps/qq/repositories/GuestSessionElasticRepository.kt
@@ -1,6 +1,6 @@
 package pt.isel.ps.qq.repositories
 
 import org.springframework.data.elasticsearch.repository.ElasticsearchRepository
-import pt.isel.ps.qq.repositories.elasticdocs.GuestSessionDoc
+import pt.isel.ps.qq.data.elasticdocs.GuestSessionDoc
 
 interface GuestSessionElasticRepository: ElasticsearchRepository<GuestSessionDoc, String>
