+++ code/jvm/src/main/kotlin/pt/isel/ps/qq/repositories/SessionElasticRepository.kt
@@ -3,8 +3,8 @@ package pt.isel.ps.qq.repositories
 import org.springframework.data.elasticsearch.repository.ElasticsearchRepository
 import org.springframework.stereotype.Repository
 import pt.isel.ps.qq.repositories.customelastic.SessionCustomElasticsearchRepository
-import pt.isel.ps.qq.repositories.elasticdocs.QqStatus
-import pt.isel.ps.qq.repositories.elasticdocs.SessionDoc
+import pt.isel.ps.qq.data.elasticdocs.QqStatus
+import pt.isel.ps.qq.data.elasticdocs.SessionDoc
 
 
 @Repository
