+++ code/jvm/src/main/kotlin/pt/isel/ps/qq/repositories/UserElasticRepository.kt
@@ -2,7 +2,7 @@ package pt.isel.ps.qq.repositories
 
 import org.springframework.data.elasticsearch.repository.ElasticsearchRepository
 import org.springframework.stereotype.Repository
-import pt.isel.ps.qq.repositories.elasticdocs.UserDoc
+import pt.isel.ps.qq.data.elasticdocs.UserDoc
 
 @Repository
 interface UserElasticRepository: ElasticsearchRepository<UserDoc, String>
